docs(Accordion): Move custom title and content example to Usages
