+++ docs/app/Examples/modules/Accordion/Usage/index.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+import ComponentExample from 'docs/app/Components/ComponentDoc/ComponentExample'
+import ExampleSection from 'docs/app/Components/ComponentDoc/ExampleSection'
+
+const AccordionUsageExamples = () => (
+  <ExampleSection title='Usage'>
+    <ComponentExample
+      title='Panels Prop with custom title and content'
+      examplePath='modules/Accordion/Usage/AccordionExamplePanelsPropWithCustomTitleAndContent'
+    />
+  </ExampleSection>
+)
+
+export default AccordionUsageExamples
