+++ docs/app/Examples/modules/Accordion/index.js
@@ -1,11 +1,13 @@
 import React from 'react'
 import Types from './Types'
 import Variations from './Variations'
+import Usage from './Usage'
 
 const AccordionExamples = () => (
   <div>
     <Types />
     <Variations />
+    <Usage />
   </div>
 )
